cart  working

+++ foodOrderApplication/foodOrderApplication/src/main/java/com/example/foodOrderApplication/repository/FoodRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FoodRepository extends JpaRepository<Food,Integer> {
+public interface FoodRepository extends JpaRepository<Food,Long> {
     List<Food> findByHotelId(Long id);
 }
